feat(subscription): add experimental fields for commitment amount and overage factor in Subscription model
